Create turing test
Alice has two strings, initial and goal. She can remove some number of characters from initial, which will give her a subsequence of that string. A string with no deletions is still considered a subsequence of itself. Given these two strings, can you find the minimum number of subsequences of initial that, when appended together, will form goal?



Functions

minimumConcat() has two parameters:

initial: the source string that you will get subsequences from
goal: the target string that needs to be formed


Input Format

For some of our templates, we have handled parsing for you. If we do not provide you a parsing function, you will need to parse the input directly. In this problem, our input format is as follows:

The first line is the initial String that we will be generating subsequences from
The second line is the goal String to form

Here is an example of the raw input:

abc
bcbac



Expected Output

Return the number of minimum possible subsequences of initial that can be appended together to form goal.

If there are no possible solutions, return -1.



Constraints
Both initial and goal can only contain lowercase letters, [a - z]